# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Patient": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"age": {
						"type": "number",
						"format": "double"
					},
					"gender": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"diagnoses": {
						"items": {
							"$ref": "#/components/schemas/Diagnose"
						},
						"type": "array"
					},
					"medications": {
						"items": {
							"$ref": "#/components/schemas/Medication"
						},
						"type": "array"
					},
					"appointments": {
						"items": {
							"$ref": "#/components/schemas/Appointment"
						},
						"type": "array"
					},
					"allergies": {
						"items": {
							"$ref": "#/components/schemas/Allergy"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"age",
					"gender",
					"email",
					"phone",
					"diagnoses",
					"medications",
					"appointments",
					"allergies",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": true
			},
			"Diagnose": {
				"properties": {
					"id": {
						"type": "string"
					},
					"date": {
						"type": "string",
						"format": "date-time"
					},
					"diagnosis": {
						"type": "string"
					},
					"patient": {
						"$ref": "#/components/schemas/Patient"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"date",
					"diagnosis",
					"patient",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": true
			},
			"Medication": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"dosage": {
						"type": "string"
					},
					"frequency": {
						"type": "string"
					},
					"patient": {
						"$ref": "#/components/schemas/Patient"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"dosage",
					"frequency",
					"patient",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": true
			},
			"Appointment": {
				"properties": {
					"id": {
						"type": "string"
					},
					"date": {
						"type": "string"
					},
					"time": {
						"type": "string"
					},
					"doctor": {
						"type": "string"
					},
					"department": {
						"type": "string"
					},
					"patient": {
						"$ref": "#/components/schemas/Patient"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"date",
					"time",
					"doctor",
					"department",
					"patient",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": true
			},
			"Allergy": {
				"properties": {
					"id": {
						"type": "string"
					},
					"allergy": {
						"type": "string"
					},
					"patient": {
						"$ref": "#/components/schemas/Patient"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"allergy",
					"patient",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": true
			},
			"PatientResponse": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/Patient"
						},
						"type": "array"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"lastPage": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"total",
					"page",
					"lastPage"
				],
				"type": "object",
				"additionalProperties": true
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "patient-dashboard-api",
		"version": "0.0.1",
		"description": "Patient Dashboard API",
		"contact": {}
	},
	"paths": {
		"/patients": {
			"get": {
				"operationId": "GetPatiens",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Patient"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/patients/pagination": {
			"get": {
				"operationId": "GetPatiensPagination",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PatientResponse"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/patients/{id}": {
			"get": {
				"operationId": "GetPatient",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Patient"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/patients/{id}/diagnoses": {
			"get": {
				"operationId": "GetPatientDiagnoses",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Diagnose"
									},
									"type": "array",
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/patients/{id}/medications": {
			"get": {
				"operationId": "GetPatientMedications",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Medication"
									},
									"type": "array",
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/patients/{id}/allergies": {
			"get": {
				"operationId": "GetPatientAllergies",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Allergy"
									},
									"type": "array",
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/patients/{id}/appointments": {
			"get": {
				"operationId": "GetPatientAppointments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Appointment"
									},
									"type": "array",
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}